Rework web push demo as two stimulus controllers
