+++ app/javascript/controllers/index.js
@@ -9,10 +9,12 @@ import Darkmode from './darkmode';
 import TableOfContents from './table-of-contents';
 
 import PwaInstallation from './pwa/installation';
-import PwaWebPush from './pwa/web-push';
+import PwaWebPushSubscription from './pwa/web-push-subscription';
+import PwaWebPushDemo from './pwa/web-push-demo';
 
 application.register('clipboard-copy', ClipboardCopy);
 application.register('darkmode', Darkmode);
 application.register('table-of-contents', TableOfContents);
 application.register('pwa-installation', PwaInstallation);
-application.register('pwa-web-push', PwaWebPush);
+application.register('pwa-web-push-subscription', PwaWebPushSubscription);
+application.register('pwa-web-push-demo', PwaWebPushDemo);
