--- app/views/components/pwa/_web_push_subscribe_demo.html.erb
@@ -1,32 +1,0 @@
-<%= tag.div data: { controller: "pwa-web-push", webPushKey: web_push_key }, class: "pwa-web-push group" do %>
-  <div class="flex space-x-2">
-    <%= tag.div do %>
-      <%= button_tag "Subscribe to Push Notifications",
-        disabled: true,
-        data: {
-          action: "pwa-web-push#subscribe",
-          "pwa-web-push-target": "subscribeButton"
-        },
-        class: "subscribe button primary" %>
-    <% end %>
-
-  <%= tag.div do %>
-    <%= button_tag "Unsubscribe from Push Notifications",
-      disabled: true,
-      data: {
-        action: "pwa-web-push#unsubscribe",
-        "pwa-web-push-target": "unsubscribeButton"
-      },
-      class: "unsubscribe button primary" %>
-  <% end %>
-  </div>
-
-  <%= form_for :web_push, url: pwa_web_pushes_path, data: {"pwa-web-push-target": "sendDemoPushForm"} do |f| %>
-    <%= f.hidden_field :subscription, data: { "pwa-web-push-target": "subscriptionField" } %>
-    <%= tag.fieldset data: { "pwa-web-push-target": "sendPushDemoFieldset" }, disabled: true do %>
-      <%= f.text_field :title, value: "Joy of Rails" %>
-      <%= f.text_field :message, value: "Hello World" %>
-      <%= button_tag "Send push notification to myself", class: "button primary" %>
-    <% end %>
-  <% end %>
-<% end %>
